관리자 기능 추가

+++ public/do_delete_article.php
@@ -1,8 +1,25 @@
 <?php
-include_once 'db.php';
+include_once '../start.php';
 ?>
 <?php
 $_REQUEST['id'] = mysqli_real_escape_string($dbLink, $_REQUEST['id']);
+
+$sql = "
+SELECT * FROM article
+WHERE id = '{$_REQUEST['id']}'
+";
+$rs = mysqli_query($dbLink, $sql);
+$article = mysqli_fetch_assoc($rs);
+
+if ( App__actorCanModify($article) == false ) {
+  ?>
+  <script>
+  alert('권한이 없습니다.');
+  history.back();
+  </script>
+  <?php
+  exit;
+}
 
 $sql = "
 DELETE FROM article
